moved qt importer to util
