+++ python/tank/util/qt_importer.py
